Add files via upload
lgals3_human.txt is a fasta sequence of the lgals3 human gene. lgals3_mouse.txt is a fasta sequence of the lgals3 mouse gene. Project 2 data is an excel sheet with genes and their expression values that were recorded in female mice and different points in virgin, pregnant and retired breeding mice. Project2C.nb.html is the project 2 code rough draft.
My scientific question is during pregnancy, what genes are up- and down-regulated? Additionally, if there are significant gene expression changes during pregnancy, do they return to ‘virgin’ levels? My hypothesis is that if there are gene expression changes during pregnancy, then the genes that are down-regulated will be those related to decreasing the mother’s immune response and the ones up-regulated will be for increasing nutrient flow to the baby. Furthermore, if the mice and human genome is significantly similar, then these results from mice experiments can be translated to humans.